pyproject.toml
[tool.poetry]
name = "switchbot"
version = "0.1.0"
description = "Switchbot temperature meter web json API"
authors = ["David Francos <xayon>"]

[tool.poetry.dependencies]
python = "^3.8"
aiojobs = "^0.2.2"
cleo = "^0.8.1"
bluepy = "^1.3.0"
aioswitchbotmeter = "^0.1.0"
aiohttp_jinja2 = "^1.2.0"
sqlalchemy = "^1.3.18"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
switchbot = 'switchbot:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

